[C] Fix client image reference counting. (#1524)
* [C] Fix client image reference counting.

Noticed that image reference count in the C client sometimes had
funny values around 4294967296. Turns out that our macro with -1
literal gets compiled to:

58	    AERON_GET_AND_ADD_INT64(res, var, -1);
   0x000000000000c74b <+39>:	b8 ff ff ff ff	mov    $0xffffffff,%eax
   0x000000000000c750 <+44>:	f0 48 0f c1 45 e8	lock xadd %rax,-0x18(%rbp)

So instead of adding -1, we're adding 4294967295 (0xffffffff).

With this change, it looks like this:

51	    AERON_GET_AND_ADD_INT64(res, var, -1);
   0x000000000000c6c9 <+39>:	48 c7 c0 ff ff ff ff	mov    $0xffffffffffffffff,%rax
   0x000000000000c6d0 <+46>:	f0 48 0f c1 45 e8	lock xadd %rax,-0x18(%rbp)

* [C] Make type of literals passed to AERON_GET_AND_ADD_INT64 explicit.

Addresses Mike's PR comment.

+++ aeron-client/src/main/c/concurrent/aeron_atomic64_gcc_x86_64.h
@@ -51,7 +51,7 @@ do \
     __asm__ volatile( \
         "lock; xaddq %0, %1" \
         : "=r"(original), "+m"(dst) \
-        : "0"(value)); \
+        : "0"((int64_t)value)); \
 } \
 while (false) \
